refactor: test패키지의 vo 패키지로 CarNameTest, CarPositionTest 클래스 이동
